+++ src/__test__/components/ReferralButton.test.jsx
@@ -8,7 +8,7 @@ import '@aws-amplify/auth';
 
 import pushNotificationMessage from 'utils/pushNotificationMessage';
 import endUserMessages from 'utils/endUserMessages';
-import ReferralButton from 'components/ReferralButton';
+import ReferralButton from 'components/header/ReferralButton';
 import '__test__/test-utils/setupTests';
 
 jest.mock('@aws-amplify/auth', () => ({
