+++ src/__test__/pages/experiments/[experimentId]/plots-and-tables/volcano/index.test.jsx
@@ -34,7 +34,7 @@ const experimentId = fake.EXPERIMENT_ID;
 const plotUuid = 'volcanoPlotMain';
 const defaultProps = { experimentId };
 
-jest.mock('components/UserButton', () => () => <></>);
+jest.mock('components/header/UserButton', () => () => <></>);
 jest.mock('react-resize-detector', () => (props) => {
   // eslint-disable-next-line react/prop-types
   const { children } = props;
