+++ src/pages/401.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Result, Button, Typography } from 'antd';
 import Auth from '@aws-amplify/auth';
-import FeedbackButton from 'components/FeedbackButton';
+import FeedbackButton from 'components/header/FeedbackButton';
 
 const { Title } = Typography;
